requirements.txt
altgraph==0.17.2
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.58.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
flake8==6.1.0
future==0.18.2
idna==3.4
macholib==1.15.2
mccabe==0.7.0
oauthlib==3.2.2
pip==23.3.1
platformdirs==3.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==58.0.4
six==1.15.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
wheel==0.37.0

# npm install @react-navigation/drawer para menu lateral
# pip install Pillow para relacionado a multimedia
# pip install moviepy
# /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
    # echo 'eval $(/opt/homebrew/bin/brew shellenv)' >> ~/.zshrc
    # eval $(/opt/homebrew/bin/brew shellenv)
# brew install ffmpeg
# pip install django-ffmpeg